+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Clear History Options
   - Clear All Warnings
 - Button to Enable Real-time protection on the Indicator when its disabled
+- Confirmation before disabling the Real Time Scan
 ### Improved
 - The App Background Color to a **Lighter shade** (light mode) and **Darker shade** (dark mode)